arcface-human-checker-main/requirements.txt
streamlit
opencv-python-headless
insightface
onnxruntime
scikit-learn
joblib
Pillow